Split dir_permissions into platform-agnostic interface and Unix implementation
Move syscall.Stat_t usage and UID/GID constants into dir_permissions_unix.go
(build tag !windows) so the package compiles on non-Unix platforms. The
DirectoryPermChecker interface and path-length constants remain in the
platform-agnostic dir_permissions.go, keeping toctou.go compilable everywhere.
